Statistics test refactoring
