+++ ydb/core/statistics/aggregator/ut/ut_aggregator.cpp
@@ -7,157 +7,13 @@
 #include <ydb/core/statistics/events.h>
 #include <ydb/core/statistics/service/service.h>
 
-#include <ydb/public/sdk/cpp/client/ydb_result/result.h>
-#include <ydb/public/sdk/cpp/client/ydb_table/table.h>
-#include <ydb/public/sdk/cpp/client/ydb_scheme/scheme.h>
-
 #include <thread>
 
 namespace NKikimr {
 namespace NStat {
 
-using namespace NYdb;
-using namespace NYdb::NTable;
-using namespace NYdb::NScheme;
-
 namespace {
 
-void CreateUniformTable(TTestEnv& env, const TString& databaseName, const TString& tableName) {
-    TTableClient client(env.GetDriver());
-    auto session = client.CreateSession().GetValueSync().GetSession();
-
-    auto result = session.ExecuteSchemeQuery(Sprintf(R"(
-        CREATE TABLE `Root/%s/%s` (
-            Key Uint64,
-            Value Uint64,
-            PRIMARY KEY (Key)
-        )
-        WITH ( UNIFORM_PARTITIONS = 4 );
-    )", databaseName.c_str(), tableName.c_str())).GetValueSync();
-    UNIT_ASSERT_C(result.IsSuccess(), result.GetIssues().ToString());
-
-    TStringBuilder replace;
-    replace << Sprintf("REPLACE INTO `Root/%s/%s` (Key, Value) VALUES ",
-        databaseName.c_str(), tableName.c_str());
-    for (ui32 i = 0; i < 4; ++i) {
-        if (i > 0) {
-            replace << ", ";
-        }
-        ui64 value = 4000000000000000000ull * (i + 1);
-        replace << Sprintf("(%" PRIu64 "ul, %" PRIu64 "ul)", value, value);
-    }
-    replace << ";";
-    result = session.ExecuteDataQuery(replace, TTxControl::BeginTx().CommitTx()).GetValueSync();
-    UNIT_ASSERT_C(result.IsSuccess(), result.GetIssues().ToString());
-}
-
-void CreateColumnStoreTable(TTestEnv& env, const TString& databaseName, const TString& tableName,
-    int shardCount)
-{
-    TTableClient client(env.GetDriver());
-    auto session = client.CreateSession().GetValueSync().GetSession();
-
-    auto fullTableName = Sprintf("Root/%s/%s", databaseName.c_str(), tableName.c_str());
-    auto result = session.ExecuteSchemeQuery(Sprintf(R"(
-        CREATE TABLE `%s` (
-            Key Uint64 NOT NULL,
-            Value Uint64,
-            PRIMARY KEY (Key)
-        )
-        PARTITION BY HASH(Key)
-        WITH (
-            STORE = COLUMN,
-            AUTO_PARTITIONING_MIN_PARTITIONS_COUNT = %d
-        );
-    )", fullTableName.c_str(), shardCount)).GetValueSync();
-    UNIT_ASSERT_C(result.IsSuccess(), result.GetIssues().ToString());
-
-    NYdb::TValueBuilder rows;
-    rows.BeginList();
-    for (size_t i = 0; i < 100; ++i) {
-        auto key = TValueBuilder().Uint64(i).Build();
-        auto value = TValueBuilder().OptionalUint64(i).Build();
-        rows.AddListItem();
-        rows.BeginStruct();
-        rows.AddMember("Key", key);
-        rows.AddMember("Value", value);
-        rows.EndStruct();
-    }
-    rows.EndList();
-
-    result = client.BulkUpsert(fullTableName, rows.Build()).GetValueSync();
-    UNIT_ASSERT_C(result.IsSuccess(), result.GetIssues().ToString());
-}
-
-void DropTable(TTestEnv& env, const TString& databaseName, const TString& tableName) {
-    TTableClient client(env.GetDriver());
-    auto session = client.CreateSession().GetValueSync().GetSession();
-
-    auto result = session.ExecuteSchemeQuery(Sprintf(R"(
-        DROP TABLE `Root/%s/%s`;
-    )", databaseName.c_str(), tableName.c_str())).GetValueSync();
-    UNIT_ASSERT_C(result.IsSuccess(), result.GetIssues().ToString());
-}
-
-std::shared_ptr<TCountMinSketch> ExtractCountMin(TTestActorRuntime& runtime, TPathId pathId) {
-    auto statServiceId = NStat::MakeStatServiceID(runtime.GetNodeId(1));
-
-    NStat::TRequest req;
-    req.PathId = pathId;
-    req.ColumnTag = 1;
-
-    auto evGet = std::make_unique<TEvStatistics::TEvGetStatistics>();
-    evGet->StatType = NStat::EStatType::COUNT_MIN_SKETCH;
-    evGet->StatRequests.push_back(req);
-
-    auto sender = runtime.AllocateEdgeActor(1);
-    runtime.Send(statServiceId, sender, evGet.release(), 1, true);
-    auto evResult = runtime.GrabEdgeEventRethrow<TEvStatistics::TEvGetStatisticsResult>(sender);
-
-    UNIT_ASSERT(evResult);
-    UNIT_ASSERT(evResult->Get());
-    UNIT_ASSERT(evResult->Get()->StatResponses.size() == 1);
-
-    auto rsp = evResult->Get()->StatResponses[0];
-    auto stat = rsp.CountMinSketch;
-    UNIT_ASSERT(rsp.Success);
-    UNIT_ASSERT(stat.CountMin);
-
-    return stat.CountMin;
-}
-
-void ValidateCountMin(TTestActorRuntime& runtime, TPathId pathId) {
-    auto countMin = ExtractCountMin(runtime, pathId);
-
-    for (ui32 i = 0; i < 4; ++i) {
-        ui64 value = 4000000000000000000ull * (i + 1);
-        auto probe = countMin->Probe((const char *)&value, sizeof(ui64));
-        UNIT_ASSERT_VALUES_EQUAL(probe, 1);
-    }
-}
-
-void ValidateCountMinAbsense(TTestActorRuntime& runtime, TPathId pathId) {
-    auto statServiceId = NStat::MakeStatServiceID(runtime.GetNodeId(1));
-
-    NStat::TRequest req;
-    req.PathId = pathId;
-    req.ColumnTag = 1;
-
-    auto evGet = std::make_unique<TEvStatistics::TEvGetStatistics>();
-    evGet->StatType = NStat::EStatType::COUNT_MIN_SKETCH;
-    evGet->StatRequests.push_back(req);
-
-    auto sender = runtime.AllocateEdgeActor(1);
-    runtime.Send(statServiceId, sender, evGet.release(), 1, true);
-    auto evResult = runtime.GrabEdgeEventRethrow<TEvStatistics::TEvGetStatisticsResult>(sender);
-
-    UNIT_ASSERT(evResult);
-    UNIT_ASSERT(evResult->Get());
-    UNIT_ASSERT(evResult->Get()->StatResponses.size() == 1);
-
-    auto rsp = evResult->Get()->StatResponses[0];
-    UNIT_ASSERT(!rsp.Success);
-}
 
 } // namespace
