+++ ydb/core/statistics/ut_common/ut_common.h
@@ -60,8 +60,21 @@ void CreateDatabase(TTestEnv& env, const TString& databaseName, size_t nodeCount
 
 void CreateServerlessDatabase(TTestEnv& env, const TString& databaseName, TPathId resourcesDomainKey);
 
-TPathId ResolvePathId(TTestActorRuntime& runtime, const TString& path,
-    TPathId* domainKey = nullptr, ui64* tabletId = nullptr);
+TPathId ResolvePathId(TTestActorRuntime& runtime, const TString& path, TPathId* domainKey = nullptr, ui64* tabletId = nullptr);
+
+
+TVector<ui64> GetTableShards(TTestActorRuntime& runtime, TActorId sender, const TString &path);
+TVector<ui64> GetColumnTableShards(TTestActorRuntime& runtime, TActorId sender,const TString &path);
+
+void CreateUniformTable(TTestEnv& env, const TString& databaseName, const TString& tableName);
+void CreateColumnStoreTable(TTestEnv& env, const TString& databaseName, const TString& tableName, int shardCount);
+void DropTable(TTestEnv& env, const TString& databaseName, const TString& tableName);
+
+std::shared_ptr<TCountMinSketch> ExtractCountMin(TTestActorRuntime& runtime, TPathId pathId);
+void ValidateCountMin(TTestActorRuntime& runtime, TPathId pathId);
+void ValidateCountMinAbsense(TTestActorRuntime& runtime, TPathId pathId);
+
+void AnalyzeTable(TTestActorRuntime& runtime, ui64 tabletId, const TPathId& pathId);
 
 } // namespace NStat
 } // namespace NKikimr
